rough cleanup
