--- projects/media-viewer/src/lib/data/js-wrapper/renderOptions.model.ts
@@ -1,21 +1,0 @@
-
-export class RenderOptions {
-    documentId: string;
-    pdfDocument: any;
-    scale: any;
-    rotate: number;
-    rotationPages: {page: number, rotate: number}[];
-
-    constructor(
-        documentId: string,
-        pdfDocument: any,
-        scale: any,
-        rotate: number,
-        rotationPages: {page: number, rotate: number}[]) {
-            this.documentId = documentId;
-            this.pdfDocument = pdfDocument;
-            this.scale = scale;
-            this.rotate = rotate;
-            this.rotationPages = rotationPages;
-        }
-}
