--- projects/media-viewer/src/lib/data/pdf-render.service.ts
@@ -1,110 +1,0 @@
-import { Injectable, ElementRef } from '@angular/core';
-import { RenderOptions } from './js-wrapper/renderOptions.model';
-import { PdfWrapper } from './js-wrapper/pdf-wrapper';
-import { PdfAnnotateWrapper } from './js-wrapper/pdf-annotate-wrapper';
-import { BehaviorSubject, Subject } from 'rxjs';
-import { EmLoggerService } from '../logging/em-logger.service';
-import { RotationModel } from '../model/rotation-factory.model';
-
-@Injectable()
-export class PdfRenderService {
-
-    private viewerElementRef: ElementRef;
-    private RENDER_OPTIONS: RenderOptions;
-    private pdfPages: number;
-    private dataLoadedSubject: BehaviorSubject<boolean>;
-
-    listPagesSubject: Subject<RotationModel[]>;
-    listPages: RotationModel[] = [];
-
-    constructor(private pdfWrapper: PdfWrapper,
-                private log: EmLoggerService) {
-        this.log.setClass('PdfRenderService');
-        this.dataLoadedSubject = new BehaviorSubject(false);
-        this.listPagesSubject = new Subject();
-    }
-
-    getDataLoadedSub(): BehaviorSubject<boolean> {
-        return this.dataLoadedSubject;
-    }
-
-    dataLoadedUpdate(isLoaded: boolean) {
-        this.dataLoadedSubject.next(isLoaded);
-    }
-
-    getRenderOptions() {
-        return Object.assign({}, this.RENDER_OPTIONS);
-    }
-
-    setRenderOptions(RENDER_OPTIONS: RenderOptions): any {
-        this.RENDER_OPTIONS = RENDER_OPTIONS;
-    }
-
-    getViewerElementRef(): ElementRef {
-        return this.viewerElementRef;
-    }
-
-    getPdfPages(): number {
-        return this.pdfPages;
-    }
-
-    render(renderOptions: any) {
-
-        this.pdfWrapper.getDocument(renderOptions.url).promise.then(pdf => {
-                renderOptions.pdfDocument = pdf;
-                this.pdfPages = pdf._pdfInfo.numPages;
-                this.listPages = [];
-                const pageContainer = renderOptions.pageContainer.nativeElement;
-                pageContainer.innerHTML = '';
-                for (let i = 1; i < this.pdfPages + 1; i++) {
-                  pdf.getPage(i).then(page => {
-                      const scale = 1.0;
-                      //scale, ration, dontFlip
-                      const viewport = page.getViewport(scale, 0, false);
-                      const canvas = document.createElement('canvas');
-                      pageContainer.appendChild(canvas);
-                      const context = canvas.getContext('2d');
-                      canvas.height = viewport.viewBox[3];
-                      canvas.width = viewport.viewBox[2];
-                      const renderContext = {
-                        canvasContext: context,
-                        viewport: viewport,
-                        enableWebGL: true
-                      };
-
-                      page.render(renderContext);
-                  });
-                }
-            }).catch(
-            (error) => {
-                const errorMessage = new Error('Unable to render your supplied PDF. ' +
-                renderOptions.documentId + '. Error is: ' + error);
-                this.log.error('Encountered error while rendering the PDF:' + errorMessage);
-            });
-    }
-
-    addDomPage(pageDom: any, pageNumber: number) {
-        const pagedetails = new RotationModel(pageNumber, pageDom);
-        const index = this.listPages.findIndex(pageElement => pageElement.pageNumber === pageNumber);
-        if (index > 0) {
-            this.listPages[index] = pagedetails;
-        } else {
-            this.listPages.push(pagedetails);
-        }
-    }
-
-    addPageRotation(renderOptions: RenderOptions, pageOptions: RenderOptions, pdfPage: any): number {
-        let rotation = this.getPageRotation(pageOptions, pdfPage);
-        if (!rotation) {
-            renderOptions.rotationPages.push({page: pdfPage.pageNumber, rotate: pdfPage.rotate});
-            rotation = pdfPage.rotate;
-        }
-        return rotation;
-    }
-
-    getPageRotation(pageOptions: RenderOptions, pdfPage: any): number {
-      return pageOptions.rotationPages
-        .filter(rotateObj => rotateObj.page === pdfPage.pageNumber)
-        .map(rotateObj => rotateObj.rotate)[0];
-    }
-}
