+++ projects/media-viewer/src/lib/media-viewer.module.ts
@@ -4,27 +4,18 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { CommentsComponent } from './media-viewer/annotation/comments/comments.component';
 import { CommentItemComponent } from './media-viewer/annotation/comments/comment-item/comment-item.component';
-import { PdfAdapter } from './data/pdf-adapter';
-import { NpaService } from './data/npa.service';
-import { PdfService } from './data/pdf.service';
-import { AnnotationStoreService } from './data/annotation-store.service';
 import { PdfViewerComponent } from './media-viewer/viewers/pdf-viewer/pdf-viewer.component';
-import { Utils } from './data/utils';
-import { ApiHttpService } from './data/api-http.service';
+import { AnnotationApiHttpService } from './media-viewer/annotation/annotation-api-http.service';
 import { ContextualToolbarComponent } from './media-viewer/viewers/pdf-viewer/contextual-toolbar/contextual-toolbar.component';
-import { PdfAnnotateWrapper } from './data/js-wrapper/pdf-annotate-wrapper';
-import { PdfWrapper } from './data/js-wrapper/pdf-wrapper';
+import { PdfJsWrapper } from './media-viewer/viewers/pdf-viewer/pdf-js/pdf-js-wrapper';
 import { NgtUniversalModule } from '@ng-toolkit/universal';
 import { ImageViewerComponent } from './media-viewer/viewers/image-viewer/image-viewer.component';
 import { UnsupportedViewerComponent } from './media-viewer/viewers/unsupported-viewer/unsupported-viewer.component';
 import { MediaViewerComponent } from './media-viewer/media-viewer.component';
 import { MediaViewerService } from './media-viewer/media-viewer.service';
 import { RotationComponent } from './media-viewer/viewers/pdf-viewer/rotation-toolbar/rotation.component';
-import { RotationFactoryService } from './media-viewer/viewers/pdf-viewer/rotation-toolbar/rotation-factory.service';
-import { PdfRenderService } from './data/pdf-render.service';
 import { EmLoggerService } from './logging/em-logger.service';
 import {MediaViewerMessageService} from './media-viewer/service/media-viewer-message.service';
-import {RotateDirection, RotateOperation} from './media-viewer/service/media-viewer-message.model';
 
 @NgModule({
   imports: [
@@ -50,17 +41,9 @@ import {RotateDirection, RotateOperation} from './media-viewer/service/media-vie
     RotationComponent
 ],
   providers: [
-    PdfAnnotateWrapper,
-    PdfWrapper,
-    PdfService,
-    AnnotationStoreService,
-    PdfAdapter,
-    NpaService,
-    ApiHttpService,
-    Utils,
+    PdfJsWrapper,
+    AnnotationApiHttpService,
     MediaViewerService,
-    RotationFactoryService,
-    PdfRenderService,
     EmLoggerService,
     MediaViewerMessageService
   ],
