+++ projects/media-viewer/src/lib/media-viewer/annotation/annotation-api-http.service.ts
@@ -4,19 +4,18 @@ import { tap } from 'rxjs/operators';
 import { isPlatformServer } from '@angular/common';
 import {HttpResponse, HttpClient} from '@angular/common/http';
 import {Injectable, PLATFORM_ID, Inject} from '@angular/core';
-import {IDocumentTask} from './document-task.model';
 import {Annotation, IAnnotation, IAnnotationSet} from './annotation-set.model';
-import { EmLoggerService } from '../logging/em-logger.service';
+import { EmLoggerService } from '../../logging/em-logger.service';
 
 @Injectable()
-export class ApiHttpService {
+export class AnnotationApiHttpService {
     private baseUrl: string;
 
     constructor(private log: EmLoggerService,
                 private httpClient: HttpClient,
                 @Inject(PLATFORM_ID) private platformId,
                 private transferState: TransferState) {
-        log.setClass('ApiHttpService');
+        log.setClass('AnnotationApiHttpService');
     }
 
     setBaseUrl(baseUrl: string) {
@@ -49,15 +48,6 @@ export class ApiHttpService {
                         })
                     );
         }
-    }
-    documentTask(dmDocumentId, outputDmDocumentId, baseUrl: string): Observable<HttpResponse<IDocumentTask>> {
-        const url = `${baseUrl}/api/em-npa/document-tasks`;
-        const documentTasks = {
-            inputDocumentId: dmDocumentId,
-            outputDocumentId: outputDmDocumentId
-        };
-        this.log.info('Calling NPA media-viewer-message.service.ts-' + dmDocumentId);
-        return this.httpClient.post<IDocumentTask>(url, documentTasks, {observe: 'response'});
     }
 
     deleteAnnotation(annotation: Annotation): Observable<HttpResponse<IAnnotation>> {
