+++ projects/media-viewer/src/lib/media-viewer/annotation/annotation-set.model.ts
