+++ projects/media-viewer/src/lib/media-viewer/viewers/pdf-viewer/contextual-toolbar/contextual-toolbar.component.ts
@@ -1,10 +1,8 @@
 import {Component, OnInit, ChangeDetectorRef, OnDestroy, Output, EventEmitter} from '@angular/core';
 import { Subscription } from 'rxjs';
-import { Annotation, Comment } from '../../../../data/annotation-set.model';
-import { AnnotationStoreService } from '../../../../data/annotation-store.service';
-import { PdfService } from '../../../../data/pdf.service';
-import { EmLoggerService } from '../../../../logging/em-logger.service';
-import { PdfRenderService } from '../../../../data/pdf-render.service';
+import { Annotation, Comment } from '../../../annotation/annotation-set.model';
+import { AnnotationStoreService } from '../../../annotation/annotation-store.service';
+import {EmLoggerService} from '../../../../logging/em-logger.service';
 
 @Component({
   selector: 'app-contextual-toolbar',
@@ -23,8 +21,6 @@ export class ContextualToolbarComponent implements OnInit, OnDestroy {
 
   constructor(private annotationStoreService: AnnotationStoreService,
               private ref: ChangeDetectorRef,
-              private pdfRenderService: PdfRenderService,
-              private pdfService: PdfService,
               private log: EmLoggerService) {
     this.log.setClass('ContextualToolbarComponent');
     this.toolPos = {
@@ -65,22 +61,23 @@ export class ContextualToolbarComponent implements OnInit, OnDestroy {
   }
 
   getRelativePosition(annotationId: string): {left: number; top: number} {
-    const svgSelector = this.pdfRenderService.getViewerElementRef().nativeElement
-                          .querySelector(`g[data-pdf-annotate-id="${annotationId}"]`);
-    const highlightRect = <DOMRect>svgSelector.getBoundingClientRect();
-
-    const wrapper = this.pdfService.getAnnotationWrapper().nativeElement;
-    const wrapperRect = <DOMRect>wrapper.getBoundingClientRect();
-
-    const left = ((highlightRect.x - wrapperRect.left)
-      - 175) + highlightRect.width / 2; // Minus half the toolbar width + half the length of the highlight
-    const top = ((highlightRect.y - wrapperRect.top)
-      - 59) - 5; // Minus height of toolbar + 5px
-
-    return {
-      left,
-      top
-    };
+    // const svgSelector = this.pdfRenderService.getViewerElementRef().nativeElement
+    //                       .querySelector(`g[data-pdf-annotate-id="${annotationId}"]`);
+    // const highlightRect = <DOMRect>svgSelector.getBoundingClientRect();
+    //
+    // const wrapper = this.pdfService.getAnnotationWrapper().nativeElement;
+    // const wrapperRect = <DOMRect>wrapper.getBoundingClientRect();
+    //
+    // const left = ((highlightRect.x - wrapperRect.left)
+    //   - 175) + highlightRect.width / 2; // Minus half the toolbar width + half the length of the highlight
+    // const top = ((highlightRect.y - wrapperRect.top)
+    //   - 59) - 5; // Minus height of toolbar + 5px
+    //
+    // return {
+    //   left,
+    //   top
+    // };
+    return null;
   }
 
   hideToolBar() {
