+++ projects/media-viewer/src/lib/media-viewer/viewers/pdf-viewer/pdf-js/pdf-js-wrapper.ts
@@ -4,7 +4,7 @@ import * as pdfjsViewer from 'pdfjs-dist/web/pdf_viewer';
 import 'pdfjs-dist/web/pdf_viewer.css';
 import 'pdfjs-dist/build/pdf.worker';
 
-export class PdfWrapper {
+export class PdfJsWrapper {
 
     getDocument(documentId): any {
         return pdfjsLib.getDocument(documentId);
