+++ projects/media-viewer/src/lib/media-viewer/viewers/pdf-viewer/pdf-viewer.component.ts
@@ -1,6 +1,6 @@
 import {AfterViewInit, Component, ElementRef, Input, OnDestroy, ViewChild} from '@angular/core';
 import {MediaViewerFeatures} from '../../media-viewer-features';
-import {PdfWrapper} from '../../../data/js-wrapper/pdf-wrapper';
+import {PdfJsWrapper} from './pdf-js/pdf-js-wrapper';
 import {MediaViewerMessageService} from '../../service/media-viewer-message.service';
 import {Subscription} from 'rxjs';
 import {
@@ -28,7 +28,7 @@ export class PdfViewerComponent implements AfterViewInit, OnDestroy {
 
   subscription: Subscription;
 
-  constructor(private pdfWrapper: PdfWrapper, private mediaViewerMessageService: MediaViewerMessageService) {
+  constructor(private pdfWrapper: PdfJsWrapper, private mediaViewerMessageService: MediaViewerMessageService) {
     this.subscription = this.mediaViewerMessageService.getMessage().subscribe( newMessage =>
       this.handleMessage(newMessage) );
   }
