--- projects/media-viewer/src/lib/model/rotation-factory.model.ts
@@ -1,9 +1,0 @@
-export interface IRotationModel {
-    pageNumber: number;
-    pageDom: any;
-}
-
-export class RotationModel implements IRotationModel {
-    constructor(public pageNumber: number,
-                public pageDom: any) {}
-}
